refactor: 충전 기능 분리 및 송금 기능 삭제 - 메인 계좌에 해당하는 역할만 하도록 수정   - 송금은 거래에 해당한다.
